# openapi.json
{"openapi":"3.1.0","info":{"title":"API Edu Predict MSLQ","description":"API para predicción de desempeño académico basada en MSLQ","version":"1.0.0","x-logo":{"url":"https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"}},"paths":{"/api/v1/health":{"get":{"tags":["predictions","health"],"summary":"Health Check","description":"Verifica el estado de la API","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/predict":{"post":{"tags":["predictions"],"summary":"Realizar predicción individual","description":"Realiza una predicción de desempeño académico para una semana específica","operationId":"predict_api_v1_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/predict-student":{"post":{"tags":["predictions"],"summary":"Predicción para estudiante específico","description":"Realiza predicción de desempeño para un estudiante identificado","operationId":"predict_student_api_v1_predict_student_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentPredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentPredictionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/predict-batch":{"post":{"tags":["predictions"],"summary":"Predicción en lote","description":"Realiza predicciones para múltiples estudiantes simultáneamente","operationId":"predict_batch_api_v1_predict_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPredictionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/info":{"get":{"tags":["predictions"],"summary":"Información de la API","description":"Obtiene información sobre los modelos y características disponibles","operationId":"api_info_api_v1_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["root"],"summary":"Root","description":"Endpoint raíz","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BatchPredictionRequest":{"properties":{"week":{"type":"integer","maximum":12.0,"minimum":4.0,"title":"Week","description":"Semana de predicción"},"predictions":{"items":{"$ref":"#/components/schemas/StudentPredictionRequest"},"type":"array","title":"Predictions","description":"Lista de estudiantes para predicción"}},"type":"object","required":["week","predictions"],"title":"BatchPredictionRequest","description":"Esquema para predicción en lote"},"BatchPredictionResponse":{"properties":{"total_predictions":{"type":"integer","title":"Total Predictions"},"predictions":{"items":{"$ref":"#/components/schemas/StudentPredictionResponse"},"type":"array","title":"Predictions"},"processing_time":{"type":"number","title":"Processing Time","description":"Tiempo de procesamiento en segundos"}},"type":"object","required":["total_predictions","predictions","processing_time"],"title":"BatchPredictionResponse","description":"Esquema para respuesta de predicción en lote","example":{"predictions":[{"confidence_score":0.92,"performance_level":"Alto desempeño","performance_range":"4.1 - 5.0","prediction":2,"prediction_date":"2024-01-15T10:30:00","student_name":"Juan Pérez","week":4}],"processing_time":0.234,"total_predictions":2}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictionRequest":{"properties":{"week":{"type":"integer","maximum":12.0,"minimum":4.0,"title":"Week","description":"Semana de predicción (4, 8 o 12)"},"practica_1":{"type":"number","title":"Practica 1","description":"Calificación práctica 1"},"practica_2":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Practica 2","description":"Calificación práctica 2"},"practica_3":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Practica 3","description":"Calificación práctica 3"},"I11":{"type":"number","title":"I11"},"I18":{"type":"number","title":"I18"},"I24":{"type":"number","title":"I24"},"I33":{"type":"number","title":"I33"},"I38":{"type":"number","title":"I38"},"I40":{"type":"number","title":"I40"},"I49":{"type":"number","title":"I49"},"I50":{"type":"number","title":"I50"},"I61":{"type":"number","title":"I61"},"I63":{"type":"number","title":"I63"},"I69":{"type":"number","title":"I69"},"I70":{"type":"number","title":"I70"},"I71":{"type":"number","title":"I71"},"I75":{"type":"number","title":"I75"}},"type":"object","required":["week","practica_1","I11","I18","I24","I33","I38","I40","I49","I50","I61","I63","I69","I70","I71","I75"],"title":"PredictionRequest","description":"Esquema para solicitud de predicción individual","example":{"I11":3.5,"I18":4.0,"I24":3.8,"I33":4.1,"I38":3.9,"I40":4.2,"I49":3.7,"I50":4.0,"I61":3.6,"I63":3.9,"I69":4.1,"I70":4.0,"I71":3.8,"I75":4.2,"practica_1":4.2,"week":4}},"PredictionResponse":{"properties":{"prediction":{"type":"integer","title":"Prediction","description":"Predicción de desempeño (0: Bajo, 1: Medio, 2: Alto)"},"performance_level":{"type":"string","title":"Performance Level","description":"Descripción del nivel de desempeño"},"performance_range":{"type":"string","title":"Performance Range","description":"Rango de calificación"},"week":{"type":"integer","title":"Week","description":"Semana de predicción"},"confidence_score":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence Score","description":"Confianza de la predicción"}},"type":"object","required":["prediction","performance_level","performance_range","week","confidence_score"],"title":"PredictionResponse","description":"Esquema para respuesta de predicción","example":{"confidence_score":0.85,"performance_level":"Alto desempeño","performance_range":"4.1 - 5.0","prediction":2,"week":4}},"StudentPredictionRequest":{"properties":{"student_name":{"type":"string","title":"Student Name","description":"Nombre del estudiante"},"week":{"type":"integer","maximum":12.0,"minimum":4.0,"title":"Week","description":"Semana de predicción"},"corte_1":{"type":"number","title":"Corte 1","description":"Calificación corte 1"},"corte_2":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Corte 2","description":"Calificación corte 2"},"corte_3":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Corte 3","description":"Calificación corte 3"},"I11":{"type":"number","title":"I11"},"I18":{"type":"number","title":"I18"},"I24":{"type":"number","title":"I24"},"I33":{"type":"number","title":"I33"},"I38":{"type":"number","title":"I38"},"I40":{"type":"number","title":"I40"},"I49":{"type":"number","title":"I49"},"I50":{"type":"number","title":"I50"},"I61":{"type":"number","title":"I61"},"I63":{"type":"number","title":"I63"},"I69":{"type":"number","title":"I69"},"I70":{"type":"number","title":"I70"},"I71":{"type":"number","title":"I71"},"I75":{"type":"number","title":"I75"}},"type":"object","required":["student_name","week","corte_1","I11","I18","I24","I33","I38","I40","I49","I50","I61","I63","I69","I70","I71","I75"],"title":"StudentPredictionRequest","description":"Esquema para predicción de estudiante específico"},"StudentPredictionResponse":{"properties":{"student_name":{"type":"string","title":"Student Name"},"prediction":{"type":"integer","title":"Prediction"},"performance_level":{"type":"string","title":"Performance Level"},"performance_range":{"type":"string","title":"Performance Range"},"week":{"type":"integer","title":"Week"},"confidence_score":{"type":"number","title":"Confidence Score"},"prediction_date":{"type":"string","title":"Prediction Date"}},"type":"object","required":["student_name","prediction","performance_level","performance_range","week","confidence_score","prediction_date"],"title":"StudentPredictionResponse","description":"Esquema para respuesta de predicción de estudiante","example":{"confidence_score":0.92,"performance_level":"Alto desempeño","performance_range":"4.1 - 5.0","prediction":2,"prediction_date":"2024-01-15T10:30:00","student_name":"Juan Pérez","week":4}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}